+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/persistence/UserDAO.java
@@ -114,7 +114,7 @@ public class UserDAO extends BaseDAO {
         return retrieveObjects(query, namedParameter, parameter);
     }
 
-    public Long count(String query) throws DAOException {
+    public Integer count(String query) throws DAOException {
         return retrieveAmount(query);
     }
 }
