+++ Kitodo/src/main/java/org/kitodo/services/data/UserService.java
@@ -20,9 +20,7 @@ import de.sub.goobi.helper.ldap.Ldap;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -174,7 +172,7 @@ public class UserService extends SearchService<User> {
         return userDAO.search(query, namedParameter, parameter);
     }
 
-    public Long count(String query) throws DAOException {
+    public Integer count(String query) throws DAOException {
         return userDAO.count(query);
     }
 
@@ -561,7 +559,8 @@ public class UserService extends SearchService<User> {
     /**
      * Adds a new filter to list.
      *
-     * @param user object
+     * @param user
+     *            object
      * @param filter
      *            the filter to add
      */
@@ -579,7 +578,8 @@ public class UserService extends SearchService<User> {
     /**
      * Removes filter from list.
      *
-     * @param user objec
+     * @param user
+     *            objec
      * @param filter
      *            the filter to remove
      */
@@ -620,8 +620,7 @@ public class UserService extends SearchService<User> {
      * @param filter
      *            String
      */
-    private void addFilterToUser(User user, String filter)
-            throws CustomResponseException, DAOException, IOException {
+    private void addFilterToUser(User user, String filter) throws CustomResponseException, DAOException, IOException {
         LocalDateTime localDateTime = new LocalDateTime();
         Property property = new Property();
         property.setTitle("_filter");
